--- src/main/kotlin/io/acari/n7/theme/ThemeDefaults.kt
@@ -1,9 +1,0 @@
-package io.acari.n7.theme
-
-import javax.swing.plaf.ColorUIResource
-
-object ThemeDefaults {
-  val secondaryColor = ColorUIResource(0x000000)
-  val primaryColor = ColorUIResource(0xFFFFFF)
-  val contrailColor = ColorUIResource(0x64D7EF)
-}
