--- src/main/kotlin/io/acari/n7/ui/NormandyPositionData.kt
@@ -1,7 +1,0 @@
-package io.acari.n7.ui
-
-data class NormandyPositionData(
-  val startingX: Float,
-  val lengthOfContrail: Float,
-  val distanceBetweenCitadelAndNormandy: Int
-)
