--- src/main/kotlin/io/acari/n7/util/extensions.kt
@@ -1,11 +1,0 @@
-package io.acari.n7.util
-
-import com.intellij.ui.ColorUtil
-import java.awt.Color
-import java.util.*
-
-fun <T> T?.toOptional(): Optional<T> = Optional.ofNullable(this)
-
-fun Color.toHexString() = "#${ColorUtil.toHex(this)}"
-
-fun String.toColor() = ColorUtil.fromHex(this)
