+++ src/main/kotlin/io/unthrottled/n7/icon/NormandyIconComponent.kt
@@ -1,0 +1,8 @@
+package io.unthrottled.n7.icon
+
+import com.intellij.openapi.util.IconLoader
+
+object NormandyIconComponent {
+  fun getNormandyIcon() = IconLoader.getIcon("/normandy.svg")
+  fun getNormandyToCitadelIcon() = IconLoader.getIcon("/normandyToCitadel.svg")
+}
