+++ src/main/kotlin/io/unthrottled/n7/integration/AppLifecycleSubscriber.kt
@@ -1,0 +1,9 @@
+package io.unthrottled.n7.integration
+
+import com.intellij.ide.AppLifecycleListener
+import com.intellij.openapi.project.Project
+
+class AppLifecycleSubscriber(private val fundy: () -> Unit) : AppLifecycleListener {
+
+  override fun appStarting(projectFromCommandLine: Project?) = fundy()
+}
