+++ src/main/kotlin/io/unthrottled/n7/theme/ThemeDefaults.kt
@@ -1,0 +1,9 @@
+package io.unthrottled.n7.theme
+
+import javax.swing.plaf.ColorUIResource
+
+object ThemeDefaults {
+  val secondaryColor = ColorUIResource(0x000000)
+  val primaryColor = ColorUIResource(0xFFFFFF)
+  val contrailColor = ColorUIResource(0x64D7EF)
+}
