+++ src/main/kotlin/io/unthrottled/n7/ui/NormandyPositionData.kt
@@ -1,0 +1,7 @@
+package io.unthrottled.n7.ui
+
+data class NormandyPositionData(
+  val startingX: Float,
+  val lengthOfContrail: Float,
+  val distanceBetweenCitadelAndNormandy: Int
+)
